fix(profiler): make wire conversion cross-platform to unblock Windows release
The cooperative sampler (compiled on all platforms) imported `to_pyspy_traces`
from the `#[cfg(unix)]` `helper_client` module, so the Windows release build
failed at compile time with E0432 (unresolved import) — the v0.15.0 Release
workflow went red on both win32 targets while macOS/Linux passed.

Extract the shared wire-format → py_spy conversion into a new ungated
`profiler::wire` module (both the cooperative and elevated-helper paths emit
the same `TraceData` wire shape), and rename the now helper-launch-only file to
`helper_client/launch.rs`. Also handle `py_spy::StackTrace.pid` being `i32` on
Unix but `u32` on Windows — a second, latent error the Release never reached
because it died at the import first.

Verified with `cargo check --release --target x86_64-pc-windows-msvc --bin
basilisk` (the exact failing Release command).
